[mythtv-cmyth] Dedicated control connection for recording playback
This creates a dedicated control connection for recording playback.
The current implementation will always request the stream from the
master backend.

Future implementations should adapt to MythTV's ProgramInfo::GetPlaybackURL
and first try to read the file locally and as a fallback play it from
the slave it was recorded on.