[cmyth] Improve separator checking in cmyth_rcv_string
This improves the checking for string separators []:[].
In case a string ends with a part of a separator, the
code did not find the correct separator.

It was reported by a user that a recording "Title ["
breaks the loading of recordings because it does not
recognize the string's end. The protocol showed
"...[]:[]Title[[]:[]Subtitle...", and cmyth_rcv_string
returned "Title[[]:[]Subtitle".