[mythtv-cmyth] Use SetChannel
Prefer SetChannel instead of reopening the Live TV stream if possible.
We try to reuse our assigned active recorder if this is possible for the
new channel (channel available and not locked to a different multiplex).
The old method of reopening the whole chain is kept as fall back.
This should speed up the channel switching a bit, because we avoid rechecking
all inputs.