[mythtv-cmyth] Fix memory leak in MythDatabase::GetChannelGroups
The function cmyth_mysql_get_channelids_in_group writes the list of
channel ids for a channel group into an array of unsigned int.
The allocation is done using ref_alloc.
We have to call ref_release also for a channelCount of 0 because
ref_alloc will also allocate memory when called with size parameter of 0.